requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
httpx==0.24.1
apache-airflow==2.7.0
psycopg2-binary==2.9.7
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.27.0
